% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdataToJS.R
\name{RdataToJS}
\alias{RdataToJS}
\title{<RdataToJS>}
\usage{
RdataToJS(x)
}
\arguments{
\item{x}{unn dataframe de dos columnas.
La primera con el códligo de homologación de los datos.
La segunda con los datos.}
}
\description{
<Convierte un dataframe de R a una lista interpretable por rexportin.js>
}
