% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_base.R
\name{rhm_base}
\alias{rhm_base}
\title{<Rhm_Base>}
\usage{
rhm_base(mapjson, datos, jscode, mapname = jscode, title = "MAPA",
  popcolor = "#a4edba", mapcolor = "#770eff", dataname = "Data",
  popname = jscode, popvalue = "value", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{mapjson}{Una lista organizada en una FeatureColection frecuentemente proveniente de un mapa "json".}

\item{datos}{Un data.frame de dos columnas. En la primera de las cuales estó el código de homologación de los datos. 
Este código debe de estar contenido en las propiedades de las Features en mapjson. La segunda columna contiene
los datos que se asignan a cada feature.}

\item{jscode}{Es el nombre de la propiedad de las features con las que se homologa el código en datos.}

\item{mapname}{Nombre que se mostrara en los mapas.}

\item{title}{Titulo del mapa.}

\item{popcolor}{Color para resaltar mapa, en valor hexadecimal.}

\item{mapcolor}{Color general de mapa, en valor hexadecimal.}

\item{dataname}{Nombre de la serie de datos.}

\item{popvalue}{Dato para mostrar en el pop.}

\item{width}{Ancho.}

\item{height}{Alto.}

\item{elementId}{Identificador del elemento.}

\item{porname}{Nombre de los datos mostrados.}
}
\description{
<Crea un hihgmap base>
}
